change method syntax.
`foo.bar` means get 'bar' attribute `foo`.
Not `foo.bar()`.

+++ src/pvip_string.c
@@ -72,7 +72,7 @@ PVIP_BOOL PVIP_string_vprintf(PVIPString *str, const char*format, va_list ap) {
     va_list copied;
     va_copy(copied, ap); // copy original va_list
 
-    int size = vsnprintf(str->buf + str->len, str->buflen - str->len, format, ap);
+    size_t size = vsnprintf(str->buf + str->len, str->buflen - str->len, format, ap);
     if (size+1 > str->buflen - str->len) {
         str->buflen += size + 1;
         str->buf    = realloc(str->buf, str->buflen);
